Rather than explicitly dealing with "dependencies," allow "cross-calling"---that is, make it safe to call data miner blocks from other data miner blocks.
For example, let's say you need a (very small) table of the different residence classes that are found in a big residential survey. Before this commit, since there's no dependency management in data_miner, you would have to use the big survey file as the import source and just key it well. That's a lot of duplicate code in your residence classes model. Now you can just "cross-call" the residential survey data miner block and **proceed assuming that that table is full.**